Test:수정 - 테스트 중 오류 수정 [S08P22A101-430]

+++ ItsGrey/data/page/src/main/java/com/tntt/page/datasource/RemotePageDataSourceImpl.kt
@@ -59,7 +59,7 @@ class RemotePageDataSourceImpl @Inject constructor(
                     val documentSnapshot = querySnapshot.documents.first()
                     val id = documentSnapshot.id as String
                     val data = documentSnapshot.data
-                    val order = data?.get("order") as Int
+                    val order = (data?.get("order") as Long).toInt()
 
                     pageDto = PageDto(id, bookId, order)
                 }
